+++ typescript/ui/__tests__/integration/intergation.test.tsx
@@ -24,15 +24,15 @@ import {
   StreamCallbacks,
   ChatResponse,
 } from "@ragbits/api-client-react";
-import { useConfigContext } from "../src/core/contexts/ConfigContext/useConfigContext";
-import { ConfigContextProvider } from "../src/core/contexts/ConfigContext/ConfigContextProvider";
+import { useConfigContext } from "../../src/core/contexts/ConfigContext/useConfigContext";
+import { ConfigContextProvider } from "../../src/core/contexts/ConfigContext/ConfigContextProvider";
 import userEvent from "@testing-library/user-event";
-import PromptInput from "../src/core/components/inputs/PromptInput/PromptInput";
-import { pluginManager } from "../src/core/utils/plugins/PluginManager";
-import { ChatOptionsPlugin } from "../src/plugins/ChatOptionsPlugin";
-import { useHistoryStore } from "../src/core/stores/historyStore";
+import PromptInput from "../../src/core/components/inputs/PromptInput/PromptInput";
+import { pluginManager } from "../../src/core/utils/plugins/PluginManager";
+import { ChatOptionsPlugin } from "../../src/plugins/ChatOptionsPlugin";
+import { useHistoryStore } from "../../src/core/stores/historyStore";
+import FeedbackForm from "../../src/plugins/FeedbackPlugin/components/FeedbackForm";
 import { enableMapSet } from "immer";
-import FeedbackForm from "../src/plugins/FeedbackPlugin/components/FeedbackForm";
 
 describe("Integration tests", () => {
   enableMapSet();