+++ packages/ragbits-chat/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add persistent user settings (#719)
 - Send chat options under `user_settings` key (#721)
 - Add feedback indicator to messages, allow `extensions` in chat messages (#722)
+- Added unit tests for UI's core components (#717)
 
 ## 1.1.0 (2025-07-09)